FileInput - Apply template method pattern to eliminate duplicate code
